Added ModalFormOn
The original ModalFormOn function was designed to iterate over a series of objects and attach an "onclick" handler to all objects of that kind found in the dom. When I was using this package with the DataTables jquery package, this caused issues because datatables removes rows in pages 2+ from the dom, and the modalForm function is not triggered again when a new page is loaded. So what was occurring was that the buttons were all working on page 1, but not working at all on pages 2+.

This new function is meant to be used in the "onclick" attirbute of a button in a datatable.